Refactor TLS matcher/handler, and connection recording
This introduces some vast simplifications in how connections are
handled (I got rid of recordableConn, making Connection itself
recordable instead). We also don't remember TLS ClientHellos
anymore.

It solves a couple problems.

1)  TLS within TLS. Before, we assumed there would only be one
layer of TLS. This is not always the case: sometimes TLS is tunneled
over TLS and you want to configure a matcher on the inner TLS
connection (after terminating TLS). So, we no longer remember the
decoded ClientHello info and reuse it.

2) Recording after TLS termination. TLS termination replaces the
underlying Conn, but the recordableConn was still on the first
or inner Conn, before outer TLS was terminated. This meant that
ciphertext was recorded if using a matcher after TLS termination,
instead of the decrypted plaintext. Now, recording is done on the
current Conn handle, not always the initial one.

I probably broke some use cases here, but, hopefully not? I know
that I got some new use cases working.